+++ packages/database/src/migrations/0057_bitter_meteorite.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "projects" ADD COLUMN "default_bootstrap_ready" boolean DEFAULT false NOT NULL;